update to add methods and moved payroll detail to own repo
